requirements.txt
# Core Machine Learning Libraries
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
imageio>=2.9.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter Notebook Environment
jupyter>=1.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Utilities
tqdm>=4.60.0
h5py>=3.1.0
pathlib2>=2.3.6

# Optional: For advanced features
# tensorboard>=2.8.0  # For training visualization
# opencv-contrib-python>=4.5.0  # Extended OpenCV features 